refactor(image-controller): keep the k8s-typed watcher tests in watcher.rs
The previous split moved the whole test module out, which carried the
`k8s_openapi` imports (`JobStatus`, `ObjectMeta`) into a new path the
capability-boundary allowlist does not cover — `check-k8s-boundary.sh`
is keyed on exact file paths, and widening that documented exception set
to get green is the wrong trade.

Move only the launcher-protocol/digest tests instead. They touch no
Kubernetes API types at all, so the sibling file needs no exception, and
`watcher.rs` lands at 41876 bytes against the 51200 guard.
